Add message validation to CreateAlertCriteria
A new input field validating the presence of a message has been added to CreateAlertCriteria. The validation checks that the message field is required when creating an alert. Additionally, a new test case was added to confirm the correct functionality of this update. Also did minor refactoring in CreateAlertCriteriaImpl and CreateAlertTests.