Add team assignment feature to alerts
Implemented a new method to allow assigning teams to alerts. This update includes modifications in Alert.cs to introduce a new Task method for team assignment. Further, this change includes the addition of new models and criteria to support this feature. A set of unit tests has also been added to validate this new functionality.