Implement alert tagging and modifying functionality
This commit introduces a new ability to tag and modify tags on alerts. Alerts can now be equipped with two types of tags - 'OverwriteQuietHours' and 'Critical'. The AlertResponder concept has been swapped for AlertTags to encapsulate this new feature. The commit also contains new tests to assure the feature's functionality.