Add functionality to fetch logs and notes of alerts
The code introduces the ability to fetch the logs and notes related to specific alerts. This is done by adding two new methods to the Alert interface and corresponding implementations. The addition is complemented with respective test classes and minor adjustments to the error type class to accommodate the new functionality.